Use rubocop-rails >= 2.22.0 for better Rails 7.1 support
Before, rubocop-rails treated the Rails 7.1 `Rails.env.local?` as an
unrecognized environment name, requiring custom configuration of
`Rails/UnknownEnv` as a workaround.

Now that rubocop-rails 2.22.0 is released, this workaround is no longer
needed.

This commit removes the workaround and ensures that `rubocop-rails` is
added to the Gemfile with the `>= 2.22.0` version requirement.