Allow vite option to be used with Rails 7.2 (#83)
The Rails `--javascript` option no longer accepts "vite" as a value. We
were using this before to tell Rails to enable JavaScript, but skip any
actual framework installation, since "vite" was not a recognized value.

However in Rails 7.2, the `--javascript` option is more strictly
validated, such that "vite" is no longer allowed.

We can't use `--skip-javascript`, because then things like turbo would
also get skipped. So we have to pass _something_ as the JS framework.

Fix by passing "esbuild" instead. This tells Rails that we are using a
Node-based build step. Then when we install Vite, delete the esbuild
stuff that was added.