Implement bin/dev with run-pty instead of concurrently (#76)
The default `bin/dev` script provided by Rails for use with Vite and
other frontend build tools is based on the foreman gem. Currently, this
makes for a very poor interactive debugging experience.

With this PR nextgen will now replace the foreman-based `bin/dev` with
one that uses [run-pty](https://github.com/lydell/run-pty), which
provides a pty for each process. This makes it possible to use
interactive debuggers.

Note that the run-pty version of `bin/dev` is only applied to apps
generated with the Vite frontend option.

For more background, see:
https://mattbrictson.com/blog/better-bin-dev-script