Mute deprecations so stylelint works with overcommit (#111)
Stylelint recently introduced a deprecation that causes warnings to be
printed when using the `stylelint-declaration-strict-value` package.

```
(node:86769) [stylelint:005] DeprecationWarning: `context.fix` is being deprecated.
Please pass a `fix` callback to the `report` utility of "scale-unlimited/declaration-strict-value" instead.
```

This breaks overcommit, which tries to parse stderr from the stylelint
process to determine what files have lint issues.

To work around this problem, pass `NODE_OPTIONS=--no-deprecation` when
running stylelint via overcommit so that the warnings are not printed.