Properly identify Rails main version as 8.0.0.beta1 (#116)
Recent changes to the Rails GitHub repo caused nextgen to identify the
main branch version as "unknown". Fix how the version is determined so
that it is more reliable and doesn't rely on a fragile regex.

Now the main version is correctly identified as "8.0.0.beta1":

```
What version of Rails will you use? (Press ↑/↓ arrow to move and Enter to select)
‣ 7.2.1
  edge (7-2-stable)
  main (8.0.0.beta1)
```