Commit Gemfile.lock after adding Gemfile `ruby` declaration
Rails no longer includes a `ruby` declaration in the generated Gemfile.
When nextgen adds it, a side-effect is that the Gemfile.lock will change
the next time `bundle install` is run. This means that after generating
a project, when you run `bin/setup` for the first time, it may cause
the Gemfile.lock to change.

Fix by proactively running `bundle lock` and committing the result, so
that the Gemfile.lock in the generated project is guaranteed to be up to
date.