Fix mocha stubbing_method_on_nil deprecation warning (#146)
Fixes the following warning in nextgen-generated projects that use the
optional mocha generator:

> Mocha deprecation warning at test/support/mocha.rb:6:in 'block in <top
(required)>': `Mocha::Configuration#stubbing_method_on_nil=` is
deprecated and will be removed in a future release. `nil` is frozen in
Ruby >= v2.2 and Mocha will be dropping support for Ruby v2.1. At that
point it won't be possible to stub methods on `nil` any more.